solved with help
first approach was to remove all numbers that are sums of abundant
numbers, but this wasn't working. I looked up a solution, entered the
correct value into project euler and found another way... get the
triangle sum up to n and subtract all sums of abundant numbers from
that... this worked!